Pipeline MS testing completed

+++ utils/utils.py
@@ -1,9 +1,6 @@
 # fastapi
 from fastapi import HTTPException, status
 
-# models and schemas
-from models.pipeline import Pipeline
-
 # os and environment
 import os
 from dotenv import load_dotenv
@@ -23,48 +20,30 @@ def get_file_content(loc):
         file_content = file.read()
     return file_content
 
-def write_file(loc, file):
-    path = STORAGE_LOC + loc
-    with open(path, "wb") as buffer:
-        buffer.write(file.file.read())
+def write_file(path, file):
+    try:
+        with open(path, "wb") as buffer:
+            buffer.write(file.file.read())
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"{e}")
 
-def save_file(email, pname, file):
-
-    pipeline_loc = os.path.join(email, pname, 'pipelines', f"{file.filename}")
-
-    write_file(pipeline_loc, file)
-
-    return Pipeline(
-        email = email,
-        pname = pname,
-        name = file.filename,
-        content_loc = pipeline_loc
-    )
+def delete_file(path):
+    if os.path.exists(path):
+        os.remove(path)
+    return
 
 def update_files(email, pname, name, file):
 
     base_dir = os.path.join(STORAGE_LOC, email, pname, 'pipelines')
+
     if not os.path.exists(base_dir):
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Project not found")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Project not found in file storage")
 
     if name != file.filename:
         new_name_path = os.path.join(base_dir, f"{file.filename}")
         old_name_path = os.path.join(base_dir, f"{name}")
         os.rename(old_name_path, new_name_path)
 
-    pipeline_loc = os.path.join(email, pname, 'pipelines', f"{file.filename}")
+    pipeline_loc = STORAGE_LOC + os.path.join(email, pname, 'pipelines', f"{file.filename}")
 
     write_file(pipeline_loc, file)
-
-    return Pipeline(
-        email = email,
-        pname = pname,
-        name = file.filename,
-        content_loc=pipeline_loc,
-        content = get_file_content(pipeline_loc)
-    )
-
-def delete_files(email, pname, name):
-    path = os.path.join(STORAGE_LOC, email, pname, 'pipelines', name)
-    os.remove(path)
-    return
